pyproject.toml
[project]
name = "agent-notes"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.28.1",
  "langchain>=0.3.14",
  "langchain-anthropic>=0.3.3",
  "langchain-community>=0.3.14",
  "langchain-experimental>=0.3.4",
  "langchain-ollama>=0.2.2",
  "langchain-openai>=0.3.2",
  "langgraph>=0.2.64",
  "matplotlib>=3.10.0",
  "mcp[cli]>=1.9.4",
  "openai>=1.60.0",
  "psycopg[binary]>=3.2.6",
  "psycopg-pool>=3.2.6",
  "python-dotenv>=1.0.1",
  "tavily-python>=0.5.0",
  "psycopg2-binary>=2.9.10",
  "langchain-mcp-adapters>=0.1.9",
  "grandalf>=0.8",
]

[dependency-groups]
dev = ["notebook>=7.3.2"]

[project.scripts]
client = "mcp.client:main"
